chimera-orchestrator/requirements.txt
# Chimera Multi-Agent Orchestrator Requirements

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0

# Async HTTP
aiohttp>=3.9.0
httpx>=0.25.0

# Message Queue
redis>=5.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.0

# Optional: For enhanced functionality
websockets>=12.0